# api/docs/swagger.yaml
definitions:
...
  http.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: string
    type: object
  models.Assays:
    properties:
      item:
        items:
          $ref: '#/definitions/models.AssaysItem'
        type: array
    type: object
  models.AssaysItem:
    properties:
      barcode:
        type: string
      biomaterial_code:
        type: string
      orders:
        $ref: '#/definitions/models.AssaysOrders'
    type: object
  models.AssaysOrders:
    properties:
      item_code:
        type: string
    type: object
...
  models.CreateDirections:
    properties:
      assays:
        $ref: '#/definitions/models.Assays'
      patient:
        $ref: '#/definitions/models.Patient'
      referral:
        $ref: '#/definitions/models.Referral'
    type: object
...
  models.Field:
    properties:
      attributes:
        additionalProperties: true
        type: object
      autofill_field:
        type: string
      autofill_table:
        type: string
      default:
        type: string
      id:
        type: string
      index:
        type: string
      is_visible:
        type: boolean
      label:
        type: string
      relation_id:
        type: string
      required:
        type: boolean
      slug:
        type: string
      table_id:
        type: string
      type:
        type: string
    type: object
...
  models.Patient:
    properties:
      birth_date:
        type: string
      body_temperature:
        type: string
      document_number:
        type: string
      email:
        type: string
      first_name:
        type: string
      gender:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
      mis_id:
        type: string
      pay_category_id:
        type: string
      phone:
        type: string
    type: object
...
  models.Referral:
    properties:
      cito:
        type: string
      comment:
        type: string
      comment_1:
        type: string
      cycle_period:
        type: string
      date:
        type: string
      delivery_date:
        type: string
      department_code:
        type: string
      department_name:
        type: string
      diagnosis_code:
        type: string
      diagnosisName:
        type: string
      diuresis_ml:
        type: string
      doctor_code:
        type: string
      doctor_name:
        type: string
      ecomment:
        type: string
      height_cm:
        type: string
      last_menstruation:
        type: string
      mis_id:
        type: string
      nr:
        type: string
      pay_category_id:
        type: string
      pregnancy_week:
        type: string
      sampling_date:
        type: string
      weight_kg:
        type: string
    type: object
...
  object_builder_service.BarcodeGenerateRes:
    properties:
      number:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a api gateway
  termsOfService: https://udevs.io
paths:
...
  /v1/alfalab/directions:
    post:
      consumes:
      - application/json
      description: Create Directions
      operationId: create_patient
      parameters:
      - description: Resource-Id
        in: header
        name: Resource-Id
        required: true
        type: string
      - description: Environment-Id
        in: header
        name: Environment-Id
        required: true
        type: string
      - description: CreatePatientRequestBody
        in: body
        name: table
        required: true
        schema:
          $ref: '#/definitions/models.CreateDirections'
      produces:
      - application/json
      responses:
        "201":
          description: Field data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Field'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Create Directions
      tags:
      - AlfaLab
  /v1/alfalab/referral:
    get:
      consumes:
      - application/json
      description: Query Referral Result
      operationId: query_referral_result
      parameters:
      - description: Resource-Id
        in: header
        name: Resource-Id
        required: true
        type: string
      - description: Environment-Id
        in: header
        name: Environment-Id
        required: true
        type: string
      - description: nr
        in: query
        name: nr
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Barcode
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/object_builder_service.BarcodeGenerateRes'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: query referral result
      tags:
      - AlfaLab
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"